backend/requirements.txt
# Web Framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.5.0

# Image Processing
pillow>=10.0.0
opencv-python>=4.9.0
numpy>=1.26.0

# Machine Learning - Personal Color Analysis
pandas>=2.0.0
scikit-learn>=1.4.0
joblib>=1.3.0
tqdm>=4.66.0

# Machine Learning - Face Shape Classification (Hugging Face)
transformers>=4.40.0
torch>=2.0.0
torchvision>=0.15.0
